Refactor to use new API
We can use the new Django API to improve the fetchNetwork
action. Specifically we can determine if a network is too large
in a more efficient way. Updates also needed to be made to account
for differences in how data comes in from the new API. This includes
changes in how we can determine table names and node counts for networks.